feat(ISP-01): create initial configuration for NextJS
Create and understood the initial configuration for a NextJS project,
including the structure of pages, static files, and API routes.
