+++ src/pages/_app.tsx
@@ -1,5 +1,4 @@
-import "@/styles/globals.css";
-import type { AppProps } from "next/app";
+import { AppProps } from "next/app";
 
 export default function App({ Component, pageProps }: AppProps) {
   return <Component {...pageProps} />;
