+++ src/pages/success.tsx
@@ -1,0 +1,5 @@
+export default function Success() {
+  return (
+    <p>Success</p>
+  )
+}
